// src/app/model/outbound-payload/EmployeeDetailsPayload.ts
interface EmployeeDetailsPayload {
  AuthorizedNumber: string
  EmployeeActive: string
  FB_uid: string
  FirstName: string
  Lang: string
  LastName: string
  Operator_Desc: string
  PIN_Code: string
  TypeActive: string
  Type_Desc: string
  Type_id: string
  branchCode: string
  branchId: string
  branchName: string
  dtCreated: string
  dtUpdated: string
  idNumber: string
  mail: string
  operator_id: string
  personId: string
  phoneMain: string
  typeEnd: string
  typeStart: string
  workerCode: string
  Card_serial: string
  Card_Status: string
}

export default EmployeeDetailsPayload